Modernize code to C++17
This sets the C++ standard variable in the build to C++17, and
uses std::variant in favor of mpark::variant.

+++ test/api/exceptions/unit-invalid.cpp
@@ -28,7 +28,7 @@ TEST_CASE("invalid image", "[invalid]") {
     SECTION("source") {
         class InvalidSource : public SourceInterface {
             int64_t read(void *data, size_t length) override {
-                int available = std::min(length, buffer_.size() - read_pos_);
+                int64_t available = std::min(length, buffer_.size() - read_pos_);
                 if (available <= 0) {
                     return 0;
                 }
